June 29, We have gotten preferences and the user repositories and Room to work - a lot of issues stemmed from incorrect Manifest declarations (i.e., I did not name my application over there).
Other than that, I am a few steps closer to 0200 Notifications and I am somewhat aware of a way to change locales. I need to peruse more ways to implement android activities and improve the animation as well as the speed of the app.

Also, I need to ensure it denies information.