There was a bug in the system caused by my rampant carelessness with date communication - values were parsed and reformatted for no reason and multiple consolidate-able loops too.
I have since changed all dates to be Long - in EpochMilli.

They are only converted into formats when necessary for the UI. This is how it should have been, but c'est la vie.

Okay, that should be it. Oh, I disabled the visualization screen because I'll only have it reappear when I find more valid uses for it.

There was also a mistake in the AppContainer as it did not use the getDatabase function to build the DB and thus the destructive migration plan was not in effect.