New Bug unlocked !
Habit entry form does not work.

Also, add a dedicated "Test Data" button

And that nuke button

We have gotten to the first step of filters, but we haven't figured out how they work. I lost a lot of time debugging an issue caused by a manual SQL query and a problematic ENUM. So, just make a button and be done with it - in fact reduce the settings screen to be about these buttons. Okay, bye