Added Coil - might want to remove it in future since I have not seen it confer advantages to the project. The failing may very well be on development, but there's a bird over there, look !
I added an option to change the picture using drawables. It was a frustrating pain using Gemini for it - Googling and doc investigation is more reliable.

But, now the images can be changed from my cat's derpy face. It took a lot of effort to wrangle the solution away from the mire.

And really, it doesn't look like a complicated solution. I am not happy for having solved it - it's annoys me. But, that's what I get for leaving most of the info on the backburner and trusting Gemini to give me the best solution first-try.

It's not that bad - but I effortful wrangling is not something that keeps one un-asleep. If this were a complex problem I'd not be able to, but - ah.

Time to commit this and not think about it until we have to make a new UI and new Screens.

The current ones suffer from immense tech debt.