Alright - new bugs !
We can now collect the right data class from the two tables with proper joining - by God Joins are great !

You can add principles at the press of a button and they'd only appear today as no new records are built for yesterday and beyond.

However, this does mean at the moment that since the day before yesterday and the day before the day before yesterday are equivalent in the list bar - for some reason the filter bar does not update its date item.

I might be able to fix that with a transitory loading screen or a random filled drawable or pattern attached to empty screens.

I also have yet to add a swiping gesture to modify the date.

I also need to allow an edit feature to modify the name, description, and to allow the user to delete the principle.

The UI particulars can come in when I've finalized the app theme system. I am not happy with how static and disorganized the large majority of screens and even previews are and we could benefit a lot from the software equivalent of LTT's Set up doctor.

Because good lord are there are a lot of wires crossed !

But I am working without a design doc - so I ought to finish the functionality work and re-design it latter knowing what and how to create one and giving myself the motivation to never start an application's design without having one because it is awful.

I have made a few bars a component - shared where possible.

But I have integrated it to my liking.

There's still a very many found bug, e.g., the habit entry screen still doesn't work but I'll figure it out soon enough hopefully.

Oh, we could even open up a dialog to edit the principles instead of a new page.

But we have done a lot of work for it - and in good time. I.e., we didn't walk away every 2 minutes to "finish this really short video".

It's not seismic, nor did I perform a lot of indexing - in fact it is a bit slow and it could do with increasing the speed of retrieval + masking the delay.

Yet it is a step in the right direction.

Okay, time to commit and add a lot of issues so I can keep track of them even if I throw them to v3.