Alright - I haven't built it because I am very sleepy and not interested in fixing bugs.
But I have made myself happy with how I'm managing the weeds - and how some previews have gone successful.

I won't know if there are semantic or runtime issues until I run it, but it seems fine as I write this.

I have added a background pattern of repetitive vector images - I hope we can clip them or have them move gently side to side without much harm to performance.

We also need to find out how to colour the vectors according to material theme.

We've also modified some functions to reduce redundancy - and we have added the edit dialog.

we'll mark that issue completed once we test it.

After that, it's time for previews and Figma