Added a font, further modularized the principle screen.
The app seems slower to open on the emulator...

But my new method has already resulted in quicker debugging !

I'll fix this for the habits and the issues, etc., and get the principles stuff done and in good order soon.

If necessary, I'll gut it and store what I can without over-moduling.