IT WORKS TO SOME DEGREE.
Date traversal works by the principle's date created - so it would allow date traversal (and trigger recomposition) until the day of creation for principles. It took a long time for me to get it to work and not have it double each element because it wanted to.

Now, there are still a few things I'd like to implement and I will write them here because I am ready to sleep.

[] Sealed Interfaces to allow loading animations
[] thinner body bars
[] the action bar
[] creating the items
[] making a dedicated screen to edit the items, replete with a delete feature that effectively deletes what should have been possible with foreign key constraints but that doesn't seem to work
[] swipe gestures for the date traversal
[] renaming the two entities by details and deletes and their dot product to either the base name or have a dot suffix
[]State flow unsubscribing and modest flows that do not take up that many resources

And then do all of that for habits but add the filter item which shouldn't be too hard.

We may also benefit from not modularizing it too much, but we'll deal with that when we're done.