Added a lot more background patterns and consolidated them to one file in .ui
I added a dimension for the loading circle, and I added a loading circle. It's indeterminate and the only other thing I'd want to add is loading text for the various stages, like in Don't Starve.

I deleted the habit entry and edit ones as I'll have to revamp it anyway (copy with minor edits).

And I added a new curly font for smaller texts since Lemon was too beefy for anything smaller than 30.dp.

SingleDay is fun, and I ought to place more fonts and find something better than Ubuntu mono for the standard body text.

Oh, and I added padding to the navigation layer to avoid it interfering with the black pill gesture thing.

I ought to normalize the padding in the bars since they look like they might be off (from the top and from the bottom).

Anyways, other than pushing this stuff to the habit section, I have no complaints. The ANR's have even been reduced in my AVD !