Ported the principle changes to the habit section (mostly smoothly).
Manners and habits do have a few differences, but honestly, I'd rather aim for issues at v3.

Sure, this mothballs all work I did on the review feature, but at the same time, v3 can include more filters, user requests, more colours and indicators, and better modifications.

Oh, and if we can figure out how to remove "today" as a necessity for habits, we can improve on of the most severe penalties to any habit training regimen.

I think v3 will demand some form of modelled histories and single habit functions.

It is very much possible and even preferable as missing out on the chance to accurately track your habits is a pain worse than and on top of performance.

But we can probably survive without it until v4, issues are more relevant.